perf: per-step overhead optimizations in autoregressive generation loop
- RoPE: cache frequency bands as module buffers, pre-allocate ts buffer
- Causal mask: skip creation for T=1 streaming (pass None to SDPA)
- Module list: cache stateful modules to avoid named_modules() walk per step
- einops: replace rearrange with native view/permute (remove dependency)
- Residual connections: remove redundant .to(update) dtype casts
- Flow LM: remove redundant .to(float32) cast
- KV cache: pre-allocate capacity_indexes arange tensor
- State copy: efficient _copy_model_state replaces deepcopy

Benchmark: 3.4x -> 3.8x single worker, 7.7x -> 8.1x multi-worker

+++ pocket_tts/models/flow_lm.py
@@ -140,7 +140,8 @@ class FlowLMModel(nn.Module):
         input_ = self.input_linear(sequence)
 
         transformer_out = self.backbone(input_, text_embeddings, sequence, model_state=model_state)
-        transformer_out = transformer_out.to(torch.float32)
+        # Removed redundant .to(float32) — already float32 when running in float32 mode
+        # transformer_out = transformer_out.to(torch.float32)
         assert lsd_decode_steps > 0
 
         transformer_out = transformer_out[:, -1]
